requirements.txt
flask
psycopg2
nose
coverage
python-magic